cache dao

+++ src/Topxia/Service/System/Dao/Impl/CacheDaoImpl.php
@@ -11,26 +11,50 @@ class CacheDaoImpl extends BaseDao implements CacheDao
 
     public function getCache($id)
     {
-        $that = $this;
-
-        return $this->fetchCached("id:{$id}", $id, function ($id) use ($that) {
-            $sql = "SELECT * FROM {$that->getTable()} WHERE id = ? LIMIT 1";
-            return $that->getConnection()->fetchAssoc($sql, array($id));
+        $redis = $this->getRedis();
+        if ($redis) {
+            $key  = "{$this->table}:v{$this->getTableVersion()}:id:{$id}";
+            $data = $this->dataCached[$key];
+            if (empty($data)) {
+                $data                   = $this->getRedis()->get($key);
+                $this->dataCached[$key] = $data;
+            }
+            return $data;
+        } else {
+            $sql = "SELECT * FROM {$this->table} WHERE id = ? LIMIT 1";
+            return $this->getConnection()->fetchAssoc($sql, array($id));
         }
-
-        );
     }
 
     public function addCache($cache)
     {
-        $affected = $this->getConnection()->insert($this->table, $cache);
-        $this->clearCached();
+        $redis = $this->getRedis();
 
-        if ($affected <= 0) {
-            throw $this->createDaoException('Insert cache error.');
+        if ($redis) {
+            $key = "{$this->table}:v{$this->getTableVersion()}:name:{$cache['name']}";
+            $redis->setex($key, 2 * 60 * 60, $cache);
+            $this->dataCached[$key] = $cache;
+            return $cache;
+        } else {
+            $affected = $this->getConnection()->insert($this->table, $cache);
+
+            if ($affected <= 0) {
+                throw $this->createDaoException('Insert cache error.');
+            }
+
+            return $this->getCache($this->getConnection()->lastInsertId());
         }
+    }
 
-        return $this->getCache($this->getConnection()->lastInsertId());
+    protected function getCacheByName($name)
+    {
+        $key  = "{$this->table}:v{$this->getTableVersion()}:name:{$name}";
+        $data = $this->dataCached[$key];
+        if (empty($data)) {
+            $data                   = $this->getRedis()->get($key);
+            $this->dataCached[$key] = $data;
+        }
+        return $data;
     }
 
     public function findCachesByNames(array $names)
@@ -39,30 +63,46 @@ class CacheDaoImpl extends BaseDao implements CacheDao
             return array();
         }
 
-        $marks = str_repeat('?,', count($names) - 1).'?';
+        $redis = $this->getRedis();
 
-        $that = $this;
-        $keys = implode(',', $names);
-        return $this->fetchCached("names:{$keys}", $marks, $names, function ($marks, $names) use ($that) {
-            $sql = "SELECT * FROM {$that->getTable()} WHERE name IN ({$marks});";
-            return $that->getConnection()->fetchAll($sql, $names);
+        if ($redis) {
+            $datas = array();
+            foreach ($names as $key => $name) {
+                $datas[] = $this->getCacheByName($name);
+            }
+            return $datas;
+        } else {
+            $marks = str_repeat('?,', count($names) - 1).'?';
+
+            $sql = "SELECT * FROM {$this->getTable()} WHERE name IN ({$marks});";
+            return $this->getConnection()->fetchAll($sql, $names);
         }
-
-        );
     }
 
     public function deleteCacheByName($name)
     {
-        $result = $this->getConnection()->delete($this->table, array('name' => $name));
-        $this->clearCached();
-        return $result;
+        $redis = $this->getRedis();
+
+        if ($redis) {
+            $key = "{$this->table}:v{$this->getTableVersion()}:name:{$name}";
+            unset($this->dataCached[$key]);
+            return $redis->delete($key);
+        } else {
+            $result = $this->getConnection()->delete($this->table, array('name' => $name));
+            return $result;
+        }
     }
 
     public function deleteAllCache()
     {
-        $sql    = "DELETE FROM {$this->table}";
-        $result = $this->getConnection()->executeUpdate($sql, array());
-        $this->clearCached();
-        return $result;
+        $redis = $this->getRedis();
+
+        if ($redis) {
+            return $this->clearCached();
+        } else {
+            $sql    = "DELETE FROM {$this->table}";
+            $result = $this->getConnection()->executeUpdate($sql, array());
+            return $result;
+        }
     }
 }
